Remove any duplicated slash in bucket url
If you have duplicated slash `/` in your S3 url when syncing to it,
S3 cli will happily create those objects with the double slash.

In S3 it does not really exist the concept of folders, each object has a
key that is the full path.

Having multiple slashes in the key  causes troubles.
In the console it appears as a folder without name,
and in aws cli the output is confusing and it might look like the
directory is empty. This can be confusing for the operator.

To avoid that, we sanitise the directory before uploading.

We add a function remove_double_slashes() which will use the parameter
substitution feature ${parameter/pattern/string} from bash. In order
to ensure that all slashes are replaced, we need to iterate until no
more changes are applied.

A corresponding test is added.